+++ src/app/actions.ts
@@ -1,4 +1,5 @@
 import { Dispatch } from "@reduxjs/toolkit";
+import { LightAdjustmentProcessor } from "@shiguredo/light-adjustment";
 import { NoiseSuppressionProcessor } from "@shiguredo/noise-suppression";
 import { VirtualBackgroundProcessor } from "@shiguredo/virtual-background";
 import type { ConnectionPublisher, ConnectionSubscriber, TransportType } from "sora-js-sdk";
@@ -26,6 +27,7 @@ import {
   getBlurRadiusNumber,
   getDefaultVideoCodecType,
   getDevices,
+  getLightAdjustmentOptions,
   getMediaStreamTrackProperties,
   parseMetadata,
   parseQueryString,
@@ -205,6 +207,9 @@ export const setInitialParameter = () => {
     if (qsParams.blurRadius !== undefined) {
       dispatch(slice.actions.setBlurRadius(qsParams.blurRadius));
     }
+    if (qsParams.lightAdjustment !== undefined) {
+      dispatch(slice.actions.setLightAdjustment(qsParams.lightAdjustment));
+    }
     if (qsParams.mediaProcessorsNoiseSuppression !== undefined) {
       dispatch(slice.actions.setMediaProcessorsNoiseSuppression(qsParams.mediaProcessorsNoiseSuppression));
     }
@@ -319,6 +324,7 @@ export const copyURL = () => {
       aspectRatio: state.aspectRatio !== "" ? state.aspectRatio : undefined,
       resizeMode: state.resizeMode !== "" ? state.resizeMode : undefined,
       blurRadius: state.blurRadius !== "" ? state.blurRadius : undefined,
+      lightAdjustment: state.lightAdjustment !== "" ? state.lightAdjustment : undefined,
       multistream: state.multistream !== "" ? state.multistream : undefined,
       simulcast: state.simulcast !== "" ? state.simulcast : undefined,
       simulcastRid: state.simulcastRid !== "" ? state.simulcastRid : undefined,
@@ -409,6 +415,8 @@ type craeteMediaStreamPickedState = Pick<
   | "fakeContents"
   | "fakeVolume"
   | "frameRate"
+  | "lightAdjustment"
+  | "lightAdjustmentProcessor"
   | "mediaProcessorsNoiseSuppression"
   | "mediaType"
   | "micDevice"
@@ -591,6 +599,14 @@ async function createMediaStream(
     });
     let videoTrack = videoMediaStream.getVideoTracks()[0];
     dispatch(slice.actions.setTimelineMessage(createSoraDevtoolsMediaStreamTrackLog("start", videoTrack)));
+    if (state.lightAdjustment !== "" && LightAdjustmentProcessor.isSupported()) {
+      if (state.lightAdjustmentProcessor === null) {
+        throw new Error("Failed to start LightAdjustmentProcessor. LightAdjustmentProcessor is 'null'");
+      }
+      const options = getLightAdjustmentOptions(state.lightAdjustment);
+      state.lightAdjustmentProcessor.stopProcessing();
+      videoTrack = await state.lightAdjustmentProcessor.startProcessing(videoTrack, options);
+    }
     if (state.blurRadius !== "" && VirtualBackgroundProcessor.isSupported()) {
       if (state.virtualBackgroundProcessor === null) {
         throw new Error("Failed to start VirtualBackgroundProcessor. VirtualBackgroundProcessor is 'null'");
@@ -705,15 +721,29 @@ function setSoraCallbacks(
       message["params"] = event.params;
     }
     dispatch(slice.actions.setTimelineMessage(createSoraDevtoolsTimelineMessage("event-on-disconnect", message)));
-    const { fakeContents, soraContents, reconnect, noiseSuppressionProcessor, virtualBackgroundProcessor } = getState();
+    const {
+      fakeContents,
+      soraContents,
+      reconnect,
+      lightAdjustmentProcessor,
+      noiseSuppressionProcessor,
+      virtualBackgroundProcessor,
+    } = getState();
     const { localMediaStream, remoteMediaStreams } = soraContents;
+    let originalTrack;
+    if (lightAdjustmentProcessor && lightAdjustmentProcessor.isProcessing()) {
+      originalTrack = lightAdjustmentProcessor.getOriginalTrack();
+      lightAdjustmentProcessor.stopProcessing();
+    }
     if (virtualBackgroundProcessor && virtualBackgroundProcessor.isProcessing()) {
-      const originalTrack = virtualBackgroundProcessor.getOriginalTrack();
-      if (originalTrack) {
-        originalTrack.stop();
-        dispatch(slice.actions.setTimelineMessage(createSoraDevtoolsMediaStreamTrackLog("stop", originalTrack)));
+      if (originalTrack === undefined) {
+        originalTrack = virtualBackgroundProcessor.getOriginalTrack();
       }
       virtualBackgroundProcessor.stopProcessing();
+    }
+    if (originalTrack !== undefined) {
+      originalTrack.stop();
+      dispatch(slice.actions.setTimelineMessage(createSoraDevtoolsMediaStreamTrackLog("stop", originalTrack)));
     } else {
       if (localMediaStream) {
         localMediaStream.getVideoTracks().forEach((track) => {
@@ -722,6 +752,7 @@ function setSoraCallbacks(
         });
       }
     }
+
     if (noiseSuppressionProcessor && noiseSuppressionProcessor.isProcessing()) {
       const originalTrack = noiseSuppressionProcessor.getOriginalTrack();
       if (originalTrack) {
@@ -923,13 +954,20 @@ export const connectSora = () => {
       if (error instanceof Error) {
         dispatch(slice.actions.setSoraErrorAlertMessage(`Failed to connect Sora. ${error.message}`));
       }
+      let originalTrack;
+      if (state.lightAdjustmentProcessor && state.lightAdjustmentProcessor.isProcessing()) {
+        originalTrack = state.lightAdjustmentProcessor.getOriginalTrack();
+        state.lightAdjustmentProcessor.stopProcessing();
+      }
       if (state.virtualBackgroundProcessor && state.virtualBackgroundProcessor.isProcessing()) {
-        const originalTrack = state.virtualBackgroundProcessor.getOriginalTrack();
-        if (originalTrack) {
-          originalTrack.stop();
-          dispatch(slice.actions.setTimelineMessage(createSoraDevtoolsMediaStreamTrackLog("stop", originalTrack)));
+        if (originalTrack === undefined) {
+          originalTrack = state.virtualBackgroundProcessor.getOriginalTrack();
         }
         state.virtualBackgroundProcessor.stopProcessing();
+      }
+      if (originalTrack) {
+        originalTrack.stop();
+        dispatch(slice.actions.setTimelineMessage(createSoraDevtoolsMediaStreamTrackLog("stop", originalTrack)));
       } else {
         if (mediaStream) {
           mediaStream.getVideoTracks().forEach((track) => {
@@ -938,6 +976,7 @@ export const connectSora = () => {
           });
         }
       }
+
       if (state.noiseSuppressionProcessor && state.noiseSuppressionProcessor.isProcessing()) {
         const originalTrack = state.noiseSuppressionProcessor.getOriginalTrack();
         if (originalTrack) {
@@ -1233,6 +1272,8 @@ export const setMicDevice = (micDevice: boolean) => {
         fakeContents: state.fakeContents,
         fakeVolume: state.fakeVolume,
         frameRate: state.frameRate,
+        lightAdjustment: state.lightAdjustment,
+        lightAdjustmentProcessor: state.lightAdjustmentProcessor,
         mediaProcessorsNoiseSuppression: state.mediaProcessorsNoiseSuppression,
         mediaType: state.mediaType,
         micDevice: micDevice,
@@ -1287,6 +1328,8 @@ export const setCameraDevice = (cameraDevice: boolean) => {
         fakeContents: state.fakeContents,
         fakeVolume: state.fakeVolume,
         frameRate: state.frameRate,
+        lightAdjustment: state.lightAdjustment,
+        lightAdjustmentProcessor: state.lightAdjustmentProcessor,
         mediaProcessorsNoiseSuppression: state.mediaProcessorsNoiseSuppression,
         mediaType: state.mediaType,
         micDevice: state.micDevice,
@@ -1385,6 +1428,7 @@ export const {
   setFacingMode,
   setFrameRate,
   setIgnoreDisconnectWebSocket,
+  setLightAdjustment,
   setLocalMediaStream,
   setLogMessages,
   setLyraParamsBitrate,